feat(config, build): refactor config; build executables
* feat(config): Make changes to config to use pydantic settings

* chore: Refactor all uses of config to use the config object from the config module

* refactor(config): Remove unused code, fix typos and add comments for clearer information

* refactor: Add more comments and change variable names

* fix(record): Fix progress bar

* feature(config): Re-read config value only if it changes

* update README

---------

+++ openadapt/app/dashboard/run.py
@@ -2,38 +2,54 @@
 
 
 import os
+import pathlib
 import subprocess
+import webbrowser
 
 from loguru import logger
 
+from openadapt.build_utils import is_running_from_executable
+from openadapt.config import config
 from openadapt.extensions.thread import Thread
+
+from .api.index import run_app
 
 
 def run() -> Thread:
     """Run the dashboard web application."""
     # change to the client directory
-    cur_dir = os.path.dirname(os.path.abspath(__file__))
+    cur_dir = pathlib.Path(__file__).parent
 
     def run_client() -> subprocess.Popen:
         """The entry point for the thread that runs the dashboard client."""
-        from openadapt.config import DASHBOARD_CLIENT_PORT, DASHBOARD_SERVER_PORT
+        if is_running_from_executable():
+            webbrowser.open(
+                f"http://localhost:{config.DASHBOARD_SERVER_PORT}/recordings"
+            )
+            run_app()
+            return
 
         return subprocess.Popen(
             ["node", "index.js"],
             cwd=cur_dir,
             env={
                 **os.environ,
-                "DASHBOARD_CLIENT_PORT": str(DASHBOARD_CLIENT_PORT),
-                "DASHBOARD_SERVER_PORT": str(DASHBOARD_SERVER_PORT),
+                "DASHBOARD_CLIENT_PORT": str(config.DASHBOARD_CLIENT_PORT),
+                "DASHBOARD_SERVER_PORT": str(config.DASHBOARD_SERVER_PORT),
             },
         )
 
-    return Thread(target=run_client, daemon=True, args=())
+    return Thread(
+        target=run_client,
+        daemon=True,
+        args=(),
+    )
 
 
 def cleanup(process: subprocess.Popen) -> None:
     """Cleanup the dashboard web application process."""
     logger.debug("Terminating the dashboard client.")
-    process.terminate()
-    process.wait()
+    if process:
+        process.terminate()
+        process.wait()
     logger.debug("Dashboard client terminated.")
